+++ CHANGELOG.md
@@ -7,6 +7,7 @@ New functionality:
 * [OLMIS-6359](https://openlmis.atlassian.net/browse/OLMIS-6359): Added UI for adding and editing system notifications.
 * [OLMIS-6361](https://openlmis.atlassian.net/browse/OLMIS-6361): Added UI for displaying system notifications on Home page.
 * [OLMIS-6388](https://openlmis.atlassian.net/browse/OLMIS-6388): Adjusted UI after renaming displayed property of system notification to isDisplayed.
+* [OLMIS-6399](https://openlmis.atlassian.net/browse/OLMIS-6399): Adjusted UI after renaming openlmis.administration.systemNotification state to openlmis.administration.systemNotifications.
 
 5.5.1 / 2019-05-27
 ==================